Changed Entity file extension. Quaternions work. Added radians to degeres conversion and vice-versa.
Also added an overload for the Vector<T>.Set(size_t) method: it returns a reference to the object we're trying to change, hence you can absolutely do "myVector.Set(2) = 4". Be careful though, this method adds a new element if the current size of the vector we edit is equal to the index we pass as a parameter to the Set() method, meaning it adds an empty space with value (0) if no value is assigned after that.