Added tables according to the upcoming documentation
Serialization tables are used to store arrays and vectors in files.
That way, it's possible to store a theoretical infinite amount of data within a single table.
It's possible to create a table to write it later in a file, or retrieve a table's data from a string.

Also yes, this is part of the upcoming save feature.