Added Item, Inventory, and PlayerStats + Added Serial Tables
InventoryComponent is an attachable component to any Object that holds several Items.
An Item is an abstract idea of a single object that can be manipulated by the player or the game itself. It's the resource Inventories trade.
PlayerStats bring a simple, comprehensive and handy way to manage one's user player experience in a game. Manage a player health, stamina and even mana effortlessly without worrying about any math, already has already been done.
PlayerStats even manage a simple leveling system with event callbacks.

About Serialization
Serial Tables are a simple way to store low-level data easily and retrieve it quickly. In other words, it helps a lot in transferring arrays / vectors data from memory to the hard drive and vice versa.
I also provided a support for strings serialization following the same fashion as the Serial Tables do.